chore: rename `sage` and `monk` to `f_sage` and `f_monk`
