+++ pyproject.toml
@@ -122,12 +122,12 @@ gwk_scatter3d = "cli_gwkokab.scatter3d:main"
 genie_ecc_matters = "kokab.ecc_matters.genie:main"
 genie_n_pls_m_gs = "kokab.n_pls_m_gs.genie:main"
 ## Sage scripts
-sage_ecc_matters = "kokab.ecc_matters.sage:main"
-sage_n_pls_m_gs = "kokab.n_pls_m_gs.sage:main"
-sage_one_powerlaw_one_peak = "kokab.one_powerlaw_one_peak.sage:main"
+f_sage_ecc_matters = "kokab.ecc_matters.f_sage:main"
+f_sage_n_pls_m_gs = "kokab.n_pls_m_gs.f_sage:main"
+f_sage_one_powerlaw_one_peak = "kokab.one_powerlaw_one_peak.f_sage:main"
 ## Monk scripts
-monk_ecc_matters = "kokab.ecc_matters.monk:main"
-monk_n_pls_m_gs = "kokab.n_pls_m_gs.monk:main"
+f_monk_ecc_matters = "kokab.ecc_matters.f_monk:main"
+f_monk_n_pls_m_gs = "kokab.n_pls_m_gs.f_monk:main"
 ## PPD scripts
 ppd_ecc_matters = "kokab.ecc_matters.ppd:main"
 ppd_n_pls_m_gs = "kokab.n_pls_m_gs.ppd:main"
