+++ src/kokab/ecc_matters/f_sage.py
@@ -7,7 +7,7 @@ from typing import List
 
 from gwkokab.parameters import ECCENTRICITY, PRIMARY_MASS_SOURCE, SECONDARY_MASS_SOURCE
 from kokab.ecc_matters.common import EccentricityMattersModel
-from kokab.utils.sage import get_parser as sage_parser, Sage
+from kokab.utils.f_sage import get_parser as sage_parser, Sage
 
 
 class EccentricityMattersSage(Sage):
