+++ src/kokab/utils/__init__.py
@@ -4,13 +4,13 @@
 
 from . import (
     common as common,
+    f_monk as f_monk,
+    f_sage as f_sage,
     flowMC_based as flowMC_based,
     genie_parser as genie_parser,
     guru as guru,
-    monk as monk,
     poisson_mean_parser as poisson_mean_parser,
     ppd as ppd,
     ppd_parser as ppd_parser,
     regex as regex,
-    sage as sage,
 )
